Update `FrameId` type to `long`, enhance table styling, and improve UI consistency across components

+++ DbcViewer/Entities/DbcMessage.cs
@@ -5,7 +5,7 @@ public sealed class DbcMessage
     public Guid Id { get; set; } = Guid.NewGuid();
     public Guid DbcFileId { get; set; }
     public DbcFile DbcFile { get; set; } = null!;
-    public int FrameId { get; set; }
+    public long FrameId { get; set; }
     public string Name { get; set; } = string.Empty;
     public short LengthInBytes { get; set; }
     public string Transmitter { get; set; } = string.Empty;
